@@ -8,23 +8,25 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^8",
-        "cviebrock/laravel-elasticsearch": "^10",
+        "php": "^8.1",
         "guzzlehttp/guzzle": "^7.3",
-        "laravel/framework": "^9.0",
-        "laravel/horizon": "^5.7",
-        "laravel/tinker": "^2.0",
-        "laravel/ui": "^3.0",
+        "laravel/framework": "^10.28",
+        "laravel/horizon": "^5.12",
+        "laravel/tinker": "^2.8",
+        "laravel/ui": "^4.2",
+        "mailerlite/laravel-elasticsearch": "^11",
         "mobiledetect/mobiledetectlib": "^2.8",
+        "nunomaduro/collision": "^7.9",
         "phattarachai/laravel-mobile-detect": "^1.0",
         "squizlabs/php_codesniffer": "^3.6"
     },
     "require-dev": {
-        "barryvdh/laravel-debugbar": "^3.3",
+        "barryvdh/laravel-debugbar": "^3.8",
         "friendsofphp/php-cs-fixer": "^3.8",
-        "mockery/mockery": "^1.3.1",
-        "spatie/laravel-ignition": "^1.0",
-        "phpunit/phpunit": "^10.3"
+        "mockery/mockery": "^1.4.4",
+        "spatie/laravel-ignition": "^2.0",
+        "phpunit/phpunit": "^10.3",
+        "fakerphp/faker": "^1.9.1"
     },
     "config": {
         "optimize-autoloader": true,
@@ -53,7 +55,7 @@
             "Tests\\": "tests/"
         }
     },
-    "minimum-stability": "dev",
+    "minimum-stability": "stable",
     "prefer-stable": true,
     "scripts": {
         "post-autoload-dump": [
