@@ -1,5 +1,9 @@
 <?php
 
+use Illuminate\Support\Facades\Auth as FacAuth;
+use App\Http\Controllers\Auth;
+use App\Http\Controllers\DebugController;
+use App\Http\Controllers\LibrisController;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -17,22 +21,22 @@ use Illuminate\Support\Facades\Route;
 //     return view('welcome');
 // });
 
-Auth::routes();
+FacAuth::routes();
 
-Route::get('/auth/checkLogin', 'Auth\CheckLoginController@checkLogin');
+Route::get('/auth/checkLogin', [Auth\CheckLoginController::class, 'checkLogin']);
 
-Route::get('/', 'LibrisController@home')->name("home")->middleware('auth');
+Route::get('/', [LibrisController::class, 'home'])->name("home")->middleware('auth');
 
-Route::get('/system/{system}', 'LibrisController@docsBySystem')->name("system.index")->middleware('auth');
-Route::get('/document', 'LibrisController@showDocument')->name("document.iframe")->middleware('auth');
-Route::get('/systemList/{system}', 'LibrisController@docsBySystemList')->name("system.list")->middleware('auth');
+Route::get('/system/{system}', [LibrisController::class, 'docsBySystem'])->name("system.index")->middleware('auth');
+Route::get('/document', [LibrisController::class, 'showDocument'])->name("document.iframe")->middleware('auth');
+Route::get('/systemList/{system}', [LibrisController::class, 'docsBySystemList'])->name("system.list")->middleware('auth');
 
-Route::get('/search', 'LibrisController@search')->name("search")->middleware('auth');
+Route::get('/search', [LibrisController::class, 'search'])->name("search")->middleware('auth');
 
 
 if (App::environment(['local', 'staging'])) {
-    Route::get('/debug/thumbnail', 'DebugController@thumbnail')->name("debug.thumbnail")->middleware('auth:admin');
-    Route::get('/debug/system', 'DebugController@system')->name("debug.system")->middleware('auth:admin');
+    Route::get('/debug/thumbnail', [DebugController::class, 'thumbnail'])->name("debug.thumbnail")->middleware('auth:admin');
+    Route::get('/debug/system', [DebugController::class, 'system'])->name("debug.system")->middleware('auth:admin');
 }
 
 Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home')->middleware('auth');
