@@ -36,10 +36,8 @@ class FlushRedis extends Command
 
     /**
      * Execute the console command.
-     *
-     * @return int
      */
-    public function handle()
+    public function handle(): int
     {
         return 0;
 
