@@ -38,10 +38,8 @@ class LibrisCreate extends Command
 
     /**
      * Execute the console command.
-     *
-     * @return mixed
      */
-    public function handle(LibrisInterface $libris)
+    public function handle(LibrisInterface $libris): int
     {
            $libris->updateIndex();
            $libris->updatePipeline();
