@@ -28,12 +28,10 @@ class LibrisUpdateTags extends Command
 
     /**
      * Get the next page of results
-     *
-     * @return int
      */
 
 
-    private function nextPage($size=100)
+    private function nextPage($size=100): int
     {
         // if ($this->option('system')) {
         //     $docs = $this->libris->docsBySystem($this->option('system'), $page);
@@ -60,10 +58,8 @@ class LibrisUpdateTags extends Command
 
     /**
      * Execute the console command.
-     *
-     * @return int
      */
-    public function handle(LibrisInterface $libris)
+    public function handle(LibrisInterface $libris): int
     {
         $this->line("Updating tags, one sec...");
 
