@@ -2,6 +2,8 @@
 
 namespace App\Http\Middleware;
 
+use Symfony\Component\HttpFoundation\Response;
+use Illuminate\Http\Request;
 use App\Providers\RouteServiceProvider;
 use Closure;
 use Illuminate\Support\Facades\Auth;
@@ -12,13 +14,8 @@ class RedirectIfAuthenticated
 
     /**
      * Handle an incoming request.
-     *
-     * @param  \Illuminate\Http\Request $request
-     * @param  \Closure                 $next
-     * @param  string|null              $guard
-     * @return mixed
      */
-    public function handle($request, Closure $next, $guard=null)
+    public function handle(Request $request, Closure $next, ?string $guard=null): Response
     {
         if (Auth::guard($guard)->check()) {
             return redirect(RouteServiceProvider::HOME);
