@@ -10,10 +10,8 @@ class TrustHosts extends Middleware
 
     /**
      * Get the host patterns that should be trusted.
-     *
-     * @return array
      */
-    public function hosts()
+    public function hosts(): array
     {
         return [
             $this->allSubdomainsOfApplicationUrl(),
