@@ -20,10 +20,5 @@ class TrustProxies extends Middleware
      * @var integer
      */
     // After...
-    protected $headers =
-    Request::HEADER_X_FORWARDED_FOR |
-    Request::HEADER_X_FORWARDED_HOST |
-    Request::HEADER_X_FORWARDED_PORT |
-    Request::HEADER_X_FORWARDED_PROTO |
-    Request::HEADER_X_FORWARDED_AWS_ELB;
+    protected $headers = (Request::HEADER_X_FORWARDED_FOR | Request::HEADER_X_FORWARDED_HOST | Request::HEADER_X_FORWARDED_PORT | Request::HEADER_X_FORWARDED_PROTO | Request::HEADER_X_FORWARDED_AWS_ELB);
 }//end class
