@@ -17,7 +17,10 @@ use App\Exceptions;
 
 class ScanFile implements ShouldQueue
 {
-    use Dispatchable, InteractsWithQueue, Queueable, SerializesModels;
+    use Dispatchable;
+    use InteractsWithQueue;
+    use Queueable;
+    use SerializesModels;
 
     protected $system;
 
@@ -49,10 +52,8 @@ class ScanFile implements ShouldQueue
 
     /**
      * Job as string
-     *
-     * @return void
      */
-    public function __toString()
+    public function __toString(): string
     {
         return sprintf("ScanFile <%s>", $this->filename);
 
@@ -61,26 +62,24 @@ class ScanFile implements ShouldQueue
 
     /**
      * Execute the job.
-     *
-     * @return void
      */
-    public function handle(LibrisInterface $libris)
+    public function handle(LibrisInterface $libris): void
     {
         $this->libris = $libris;
 
-        Log::info("[Scanfile] Hello ".$this->filename);
+        Log::info("[Scanfile] Hello " . $this->filename);
 
         Redis::funnel('ScanFile')->limit(5)->then(
             function () {
-                Log::debug("[Scanfile] Got lock for ".$this->filename." = ".md5($this->filename));
+                Log::debug("[Scanfile] Got lock for " . $this->filename . " = " . md5($this->filename));
                 try {
                     $returnValue = $this->libris->addDocument($this->filename);
-                    Log::info("[Scanfile] Finished ".$this->filename);
+                    Log::info("[Scanfile] Finished " . $this->filename);
                 } catch (Exceptions\LibrisFileNotSupported $e) {
-                    Log::warning("[Scanfile] Unsupported File Type ".$this->filename);
+                    Log::warning("[Scanfile] Unsupported File Type " . $this->filename);
                     return 0;
                 } catch (Exceptions\LibrisTooLarge $e) {
-                    Log::warning("[Scanfile] File too big ".$this->filename);
+                    Log::warning("[Scanfile] File too big " . $this->filename);
                     return 0;
                 } catch (Exception $e) {
                     Log::info("[Scanfile] Caught Exception");
@@ -98,7 +97,7 @@ class ScanFile implements ShouldQueue
             function () {
                 $release = (60 + rand(0, 20));
                 // Could not obtain lock...
-                Log::debug("[Scanfile] ".$this->filename." bounced ".$release);
+                Log::debug("[Scanfile] " . $this->filename . " bounced " . $release);
                 return $this->release($release);
             }
         );
@@ -110,10 +109,8 @@ class ScanFile implements ShouldQueue
 
     /**
      * Determine the time at which the job should timeout.
-     *
-     * @return \DateTime
      */
-    public function retryUntil()
+    public function retryUntil(): DateTime
     {
         // Log::debug("[ScanFile] Add ".(60*60*24)." secs to ".$this->filename);
         return now()->addSeconds(60 * 60 * 24);
