@@ -13,10 +13,8 @@ class LibrisServiceProvider extends ServiceProvider
 
     /**
      * Register services.
-     *
-     * @return void
      */
-    public function register()
+    public function register(): void
     {
         $this->app->singleton(
             LibrisInterface::class,
@@ -30,10 +28,8 @@ class LibrisServiceProvider extends ServiceProvider
 
     /**
      * Bootstrap services.
-     *
-     * @return void
      */
-    public function boot()
+    public function boot(): void
     {
 
     }//end boot()
