@@ -10,10 +10,8 @@ class PDFBoxServiceProvider extends ServiceProvider
 
     /**
      * Register services.
-     *
-     * @return void
      */
-    public function register()
+    public function register(): void
     {
 
     }//end register()
@@ -21,10 +19,8 @@ class PDFBoxServiceProvider extends ServiceProvider
 
     /**
      * Bootstrap services.
-     *
-     * @return void
      */
-    public function boot()
+    public function boot(): void
     {
 
     }//end boot()
